fix relay by moving relaying to DiscordChatFeed
