+++ DiscordLink/Source/Systems/Config.cs
@@ -368,7 +368,6 @@ namespace Eco.Plugins.DiscordLink
         [Description("The name or ID of the Discord Server. This setting can be changed while the server is running and will in that case trigger a reconnection to Discord."), Category("Base Configuration - Discord")]
         public string DiscordServer { get; set; } = string.Empty;
         
-        [Browsable(false)]
         [Description("optional: If you want to connect to multiple Guilds, set this to override the singular DiscordServer option. Apart from being a collection, this is analogous to DiscordServer."), Category("Base Configuration - Discord")]
         public Collection<string> DiscordServers { get; set; } = new ObservableCollection<string>();
 
@@ -406,7 +405,6 @@ namespace Eco.Plugins.DiscordLink
         [Description("Discord and Eco Channels to connect together for chat crossposting. This setting can be changed while the server is running."), Category("Feeds")]
         public ObservableCollection<ChatChannelLink> ChatChannelLinks { get; set; } = new ObservableCollection<ChatChannelLink>();
 
-        [Browsable(false)]
         [Description("Two Discord Channels to connect together for chat crossposting. This setting can be changed while the server is running."), Category("Feeds")]
         public ObservableCollection<RelayChannelLink> RelayChannelLinks { get; set; } = new ObservableCollection<RelayChannelLink>();
